requirements.txt
Cython==0.19.1
Flask==0.10.1
Flask-Script==0.6.2
Jinja2==2.7.1
Mako==0.9.0
MarkupSafe==0.18
Pygments==1.6
SQLAlchemy==0.8.2
Werkzeug==0.9.4
alembic==0.6.0
amqp==1.0.13
anyjson==0.3.3
billiard==2.7.3.32
bpython==0.12
celery==3.0.23
crontab==.16
distribute==0.6.34
-e git://github.com/surfly/gevent.git@454a77ca561868854760b2d9cbfa3bf3bbd2e062#egg=gevent-dev
gevent-socketio==0.3.5-rc2
gevent-subprocess==0.1.1
gevent-websocket==0.3.6
gevent-zeromq==0.2.5
greenlet==0.4.1
ipython==1.1.0
itsdangerous==0.23
kombu==2.5.14
py==1.4.15
pytest==2.3.5
python-dateutil==2.1
pyzmq==13.1.0
redis==2.8.0
simplejson==3.3.0
six==1.4.1
wsgiref==0.1.2
